requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
python-multipart==0.0.6
jinja2==3.1.2
PyYAML>=5.4.1
mlflow>=1.20.2
pydantic==2.3.0
python-dotenv==1.0.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0
dvc==3.27.0 